test: move grace tests into controller suite
